requirements/base.txt
celery==5.3.1
django==4.2.4
django-environ==0.10.0
psycopg2-binary==2.9.7
djangorestframework==3.14.0
django-mptt==0.14.0
django-filter==23.2
Pillow==10.0.0
drf-yasg==1.21.7
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.2.2
requests==2.31.0
facebook-sdk==3.1.0
google-auth==2.22.0
stripe==6.0.0
redis==5.0.0
